pyproject.toml
[project]
name = "notion-hook"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.128.0",
    "httpx>=0.28.1",
    "pydantic-settings>=2.12.0",
    "uvicorn[standard]>=0.40.0",
]

[project.scripts]
nhook = "notion_hook:main"

[build-system]
requires = ["hatchling", "editables"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "pytest-cov>=7.0.0",
    "respx>=0.22.0",
    "ruff>=0.14.14",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
asyncio_default_fixture_loop_scope = "function"

[tool.ruff]
line-length = 88
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

